requirements.txt
ago==0.0.92
alembic==1.0.0
amqp==2.3.2
arrow==0.12.1
Babel==2.6.0
backports.functools-lru-cache==1.5
billiard==3.5.0.4
binaryornot==0.4.4
celery==4.2.1
certifi==2018.4.16
Chameleon==3.4
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
fanstatic==1.1
FormEncode==1.3.1
future==0.16.0
gevent==1.3.5
greenlet==0.4.14
gunicorn==19.9.0
hupper==1.9.1
idna==2.7
inflect==1.0.0
ipaddress==1.0.22
Jinja2==2.10
jinja2-time==0.2.0
kombu==4.2.1
linecache2==1.0.0
lingua==4.13
lxml==4.2.3
Mako==1.0.7
MarkupSafe==1.0
mysql-connector-python==8.0.11
nose==1.3.7
numpy==1.14.5
PasteDeploy==1.5.2
Pattern==3.6
Pillow==5.2.0
plaster==1.0
plaster-pastedeploy==0.6
polib==1.1.0
poyo==0.4.1
protobuf==3.6.0
pycrypto==2.6.1
Pygments==2.2.0
pypng==0.0.18
pyramid==1.10.4
pyramid-debugtoolbar==4.5
pyramid-fanstatic==0.5
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-tm==2.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
PyUtilib==5.4.1
-e git+https://github.com/qlands/pyxform.git@new_add_list_name_in_json#egg=pyxform
qrcode==6.1
qrtools==0.0.2
repoze.lru==0.7
requests==2.19.1
rutter==0.2
shutilwhich==1.1.0
six==1.11.0
SQLAlchemy==1.2.10
traceback2==1.4.0
transaction==2.2.1
translationstring==1.3
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.23
venusian==1.1.0
vine==1.1.4
waitress==1.1.0
WebHelpers==1.3
WebOb==1.8.5
whichcraft==0.4.1
xlrd==1.1.0
XlsxWriter==1.0.5
#zbar==0.10
zope.deprecation==4.3.0
zope.interface==4.5.0
zope.sqlalchemy==1.0